🔁 rename build/faucet and build/wallet to developers/faucet and developers/wallet
add redircts
