+++ packages/web/src/shared/menu-items.ts
@@ -71,11 +71,11 @@ export const menuItems = {
 export enum CeloLinks {
   agreement = '/user-agreement',
   faq = '/faq',
-  faucet = '/build/faucet',
+  faucet = '/developers/faucet',
   iconsLicense = 'https://creativecommons.org/licenses/by-nd/4.0/legalcode',
   discord = 'https://discord.gg/6yWMkgM',
   discourse = 'https://forum.celo.org/',
-  walletApp = '/build/wallet',
+  walletApp = '/developers/wallet',
   blockscout = 'http://alfajores-blockscout.celo-testnet.org/',
   disclaimer = 'https://docs.celo.org/important-information/alfajores-testnet-disclaimer',
   docs = 'https://docs.celo.org/',
